refactor: model 패키지 생성
