+++ domain-generator/src/main/kotlin/com/few/domain/generator/core/GroupNewsSummarizer.kt
@@ -1,5 +1,7 @@
 package com.few.domain.generator.core
 
+import com.few.domain.generator.core.model.GroupNewsModel
+import com.few.domain.generator.core.model.SectionContentModel
 import com.google.gson.Gson
 import com.google.gson.JsonObject
 import io.github.oshai.kotlinlogging.KotlinLogging
