+++ domain-generator/src/main/kotlin/com/few/domain/generator/core/model/NewsModel.kt
@@ -1,4 +1,4 @@
-package com.few.domain.generator.core
+package com.few.domain.generator.core.model
 
 import java.time.LocalDateTime
 import java.time.format.DateTimeFormatter
